+++ docs/changelog.md
@@ -25,6 +25,9 @@ As features stabilize some brief notes about them will accumulate here.
 * Wayland: currently being reimplemented, it maybe more unstable than usual.
   Please file GH issues for any problems you see.
   Many thanks to @tzx! #4777
+* [show_update_window](config/lua/config/show_update_window.md) has been
+  deprecated; it no longer has any effect and will be removed in a future
+  release.
 #### New
 #### Fixed
 * Race condition when very quickly adjusting font scale. Thanks to @jknockel!